Split models and code into seperate files
